Move feed to home
